split(chat.ts): refactor 1574-line handleComboChat into modular sub-services
Decompose the monolithic handleComboChat function (lines 221-1574) into
6 focused sub-modules under services/combo/chat/:

- types.ts: ExecutionContext, MutableRef<T>, TargetAttemptResult (zero-any)
- routing.ts: Init, config resolution, strategy routing, pre-screening
- auto.ts: Auto-strategy filtering, scoring, selection
- executor.ts: Per-target execution, success/error processing
- handler.ts: Set-retry loop, concurrency dispatch, response assembly
- timeout.ts: Per-model timeout wrapper (existing)

The original chat.ts is now a ~128-line thin orchestrator that sequences
routing → execution. All files under 800 lines.

Line counts: chat.ts=128, auto.ts=559, executor.ts=584, handler.ts=197,
routing.ts=168, timeout.ts=77, types.ts=79 (total=1792)

No circular dependencies. Re-exports preserve backward compatibility.
